fix: Refactor docling-eval to improve the overall codebase structure (#36)
* chore: Rename `docling/` dir as converters. Introduce `visualization/` dir.


* chore: Remove unused imports and other code formatting


* chore: Remove the `utils/` dir, delete unused files and move used code in appropriate locations


* chore: Introduce the file visualisation/visualisations.py and move there functions from benchmarks/utils.py


* chore: Update MyPy configuration in toml to override tqdm module


* chore: Clean up commented code


* chore: Add CONVERTER_TYPE and MODALITIES columns to all produced datasets


* chore: Update pinning of docling


* fix: Code refactoring:
- Move converters/teds.py into evaluators/teds.py
- Move all functions from converters/utils.py into benchmarks/utils.py.
- Rename create_xxx_converter() functions.
- Rename BenchMarkColumns.DOCLING_VERSION as BenchMarkColumns.CONVERTER_VERSION


---------
