+++ docling_eval/benchmarks/constants.py
@@ -3,7 +3,7 @@ from enum import Enum
 
 class BenchMarkColumns(str, Enum):
     CONVERTER_TYPE = "converter_type"
-    DOCLING_VERSION = "docling_version"
+    CONVERTER_VERSION = "converter_version"
     DOCLING_PIPELINE = "docling_pipeline"
 
     STATUS = "status"
