+++ docling_eval/benchmarks/cvat_annotation/eval.py
@@ -1,13 +1,4 @@
-import argparse
 import logging
-import os
-from pathlib import Path
-
-from huggingface_hub import snapshot_download
-from tabulate import tabulate  # type: ignore
-
-from docling_eval.benchmarks.constants import BenchMarkNames, EvaluationModality
-from docling_eval.cli.main import evaluate, visualise
 
 # Configure logging
 logging.basicConfig(
