+++ docling_eval/benchmarks/dpbench/create.py
@@ -1,4 +1,3 @@
-import argparse
 import json
 import logging
 import os
@@ -6,14 +5,18 @@ from pathlib import Path
 from typing import Dict, Optional
 
 from docling.datamodel.pipeline_options import TableFormerMode
-from tqdm import tqdm  # type: ignore
+from tqdm import tqdm
+
+from docling_eval.visualisation.visualisations import (  # type: ignore
+    save_comparison_html,
+    save_comparison_html_with_clusters,
+)
 
 # Configure logging
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s"
 )
 
-from bs4 import BeautifulSoup  # type: ignore
 from docling_core.types.doc.base import BoundingBox, CoordOrigin, Size
 from docling_core.types.doc.document import (
     DoclingDocument,
@@ -25,28 +28,28 @@ from docling_core.types.doc.document import (
 from docling_core.types.doc.labels import DocItemLabel
 from PIL import Image  # as PILImage
 
-from docling_eval.benchmarks.constants import BenchMarkColumns, ConverterTypes
+from docling_eval.benchmarks.constants import (
+    BenchMarkColumns,
+    ConverterTypes,
+    EvaluationModality,
+)
 from docling_eval.benchmarks.utils import (
     add_pages_to_true_doc,
     convert_html_table_into_docling_tabledata,
-    save_comparison_html,
-    save_comparison_html_with_clusters,
-    write_datasets_info,
-)
-from docling_eval.docling.conversion import (
-    create_docling_converter,
-    create_vlm_converter,
-)
-from docling_eval.docling.models.tableformer.tf_model_prediction import (
-    TableFormerUpdater,
-)
-from docling_eval.docling.utils import (
     crop_bounding_box,
     docling_version,
     extract_images,
     from_pil_to_base64uri,
     get_binary,
     save_shard_to_disk,
+    write_datasets_info,
+)
+from docling_eval.converters.conversion import (
+    create_pdf_docling_converter,
+    create_smol_docling_converter,
+)
+from docling_eval.converters.models.tableformer.tf_model_prediction import (
+    TableFormerUpdater,
 )
 
 TRUE_HTML_EXPORT_LABELS = {
@@ -247,12 +250,12 @@ def create_dpbench_e2e_dataset(
 ):
     # Create Converter
     if converter_type == ConverterTypes.DOCLING:
-        converter = create_docling_converter(page_image_scale=1.0)
+        converter = create_pdf_docling_converter(page_image_scale=1.0)
     else:
-        converter = create_vlm_converter()
+        converter = create_smol_docling_converter()
 
     # load the groundtruth
-    with open(dpbench_dir / f"dataset/reference.json", "r") as fr:
+    with open(dpbench_dir / "dataset/reference.json", "r") as fr:
         gt = json.load(fr)
 
     viz_dir = output_dir / "vizualisations"
@@ -329,7 +332,7 @@ def create_dpbench_e2e_dataset(
 
         record = {
             BenchMarkColumns.CONVERTER_TYPE: converter_type,
-            BenchMarkColumns.DOCLING_VERSION: docling_version(),
+            BenchMarkColumns.CONVERTER_VERSION: docling_version(),
             BenchMarkColumns.STATUS: str(conv_results.status),
             BenchMarkColumns.DOC_ID: str(filename),
             BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -340,6 +343,10 @@ def create_dpbench_e2e_dataset(
             BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
             BenchMarkColumns.ORIGINAL: get_binary(pdf_path),
             BenchMarkColumns.MIMETYPE: "application/pdf",
+            BenchMarkColumns.MODALITIES: [
+                EvaluationModality.LAYOUT,
+                EvaluationModality.READING_ORDER,
+            ],
         }
         records.append(record)
 
@@ -367,7 +374,7 @@ def create_dpbench_tableformer_dataset(
     tf_updater = TableFormerUpdater(mode, artifacts_path=artifacts_path)
 
     # load the groundtruth
-    with open(dpbench_dir / f"dataset/reference.json", "r") as fr:
+    with open(dpbench_dir / "dataset/reference.json", "r") as fr:
         gt = json.load(fr)
 
     viz_dir = output_dir / "vizualisations"
@@ -436,7 +443,7 @@ def create_dpbench_tableformer_dataset(
 
             record = {
                 BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: "SUCCESS",
                 BenchMarkColumns.DOC_ID: str(os.path.basename(pdf_path)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -447,6 +454,9 @@ def create_dpbench_tableformer_dataset(
                 BenchMarkColumns.GROUNDTRUTH_PICTURES: true_pictures,
                 BenchMarkColumns.PREDICTION_PAGE_IMAGES: pred_page_images,
                 BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
+                BenchMarkColumns.MODALITIES: [
+                    EvaluationModality.TABLE_STRUCTURE,
+                ],
             }
             records.append(record)
