+++ docling_eval/benchmarks/omnidocbench/create.py
@@ -1,4 +1,3 @@
-import argparse
 import glob
 import json
 import logging
@@ -6,7 +5,6 @@ import os
 from pathlib import Path
 from typing import Optional
 
-from bs4 import BeautifulSoup  # type: ignore
 from docling.datamodel.pipeline_options import TableFormerMode
 from docling_core.types.doc.base import BoundingBox, CoordOrigin, Size
 from docling_core.types.doc.document import DoclingDocument, ImageRef, ProvenanceItem
@@ -14,28 +12,32 @@ from docling_core.types.doc.labels import DocItemLabel
 from PIL import Image  # as PILImage
 from tqdm import tqdm  # type: ignore
 
-from docling_eval.benchmarks.constants import BenchMarkColumns, ConverterTypes
+from docling_eval.benchmarks.constants import (
+    BenchMarkColumns,
+    ConverterTypes,
+    EvaluationModality,
+)
 from docling_eval.benchmarks.utils import (
     add_pages_to_true_doc,
     convert_html_table_into_docling_tabledata,
-    save_comparison_html,
-    save_comparison_html_with_clusters,
-    write_datasets_info,
-)
-from docling_eval.docling.conversion import (
-    create_docling_converter,
-    create_vlm_converter,
-)
-from docling_eval.docling.models.tableformer.tf_model_prediction import (
-    TableFormerUpdater,
-)
-from docling_eval.docling.utils import (
     crop_bounding_box,
     docling_version,
     extract_images,
     from_pil_to_base64uri,
     get_binary,
     save_shard_to_disk,
+    write_datasets_info,
+)
+from docling_eval.converters.conversion import (
+    create_pdf_docling_converter,
+    create_smol_docling_converter,
+)
+from docling_eval.converters.models.tableformer.tf_model_prediction import (
+    TableFormerUpdater,
+)
+from docling_eval.visualisation.visualisations import (
+    save_comparison_html,
+    save_comparison_html_with_clusters,
 )
 
 # Configure logging
@@ -260,12 +262,12 @@ def create_omnidocbench_e2e_dataset(
 
     # Create Converter
     if converter_type == ConverterTypes.DOCLING:
-        converter = create_docling_converter(page_image_scale=1.0)
+        converter = create_pdf_docling_converter(page_image_scale=1.0)
     else:
-        converter = create_vlm_converter()
+        converter = create_smol_docling_converter()
 
     # load the groundtruth
-    with open(omnidocbench_dir / f"OmniDocBench.json", "r") as fr:
+    with open(omnidocbench_dir / "OmniDocBench.json", "r") as fr:
         gt = json.load(fr)
 
     gt = update_gt_into_map(gt)
@@ -290,7 +292,7 @@ def create_omnidocbench_e2e_dataset(
         pdf_path = Path(page_tuple[1])
 
         # logging.info(f"file: {pdf_path}")
-        if not os.path.basename(jpg_path) in gt:
+        if os.path.basename(jpg_path) not in gt:
             logging.error(f"did not find ground-truth for {os.path.basename(jpg_path)}")
             continue
 
@@ -356,7 +358,7 @@ def create_omnidocbench_e2e_dataset(
 
         record = {
             BenchMarkColumns.CONVERTER_TYPE: converter_type,
-            BenchMarkColumns.DOCLING_VERSION: docling_version(),
+            BenchMarkColumns.CONVERTER_VERSION: docling_version(),
             BenchMarkColumns.STATUS: "SUCCESS",
             BenchMarkColumns.DOC_ID: str(os.path.basename(jpg_path)),
             BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -368,6 +370,10 @@ def create_omnidocbench_e2e_dataset(
             BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
             BenchMarkColumns.GROUNDTRUTH_PAGE_IMAGES: true_page_images,
             BenchMarkColumns.GROUNDTRUTH_PICTURES: true_pictures,
+            BenchMarkColumns.MODALITIES: [
+                EvaluationModality.LAYOUT,
+                EvaluationModality.READING_ORDER,
+            ],
         }
         records.append(record)
 
@@ -395,7 +401,7 @@ def create_omnidocbench_tableformer_dataset(
     tf_updater = TableFormerUpdater(mode, artifacts_path=artifacts_path)
 
     # load the groundtruth
-    with open(omnidocbench_dir / f"OmniDocBench.json", "r") as fr:
+    with open(omnidocbench_dir / "OmniDocBench.json", "r") as fr:
         gt = json.load(fr)
 
     gt = update_gt_into_map(gt)
@@ -418,7 +424,7 @@ def create_omnidocbench_tableformer_dataset(
         pdf_path = Path(page_tuple[1])
 
         # logging.info(f"file: {pdf_path}")
-        if not os.path.basename(jpg_path) in gt:
+        if os.path.basename(jpg_path) not in gt:
             logging.error(f"did not find ground-truth for {os.path.basename(jpg_path)}")
             continue
 
@@ -474,7 +480,7 @@ def create_omnidocbench_tableformer_dataset(
 
             record = {
                 BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: "SUCCESS",
                 BenchMarkColumns.DOC_ID: str(os.path.basename(jpg_path)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -485,6 +491,7 @@ def create_omnidocbench_tableformer_dataset(
                 BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
                 BenchMarkColumns.GROUNDTRUTH_PAGE_IMAGES: true_page_images,
                 BenchMarkColumns.GROUNDTRUTH_PICTURES: true_pictures,
+                BenchMarkColumns.MODALITIES: [EvaluationModality.TABLE_STRUCTURE],
             }
             records.append(record)
