+++ docling_eval/benchmarks/tableformer_huggingface_otsl/create.py
@@ -17,21 +17,23 @@ from docling_core.types.doc.document import (
 from docling_core.types.doc.labels import DocItemLabel
 from tqdm import tqdm  # type: ignore
 
-from docling_eval.benchmarks.constants import BenchMarkColumns
+from docling_eval.benchmarks.constants import (
+    BenchMarkColumns,
+    ConverterTypes,
+    EvaluationModality,
+)
 from docling_eval.benchmarks.utils import (
     convert_html_table_into_docling_tabledata,
-    save_comparison_html,
-)
-from docling_eval.docling.models.tableformer.tf_model_prediction import (
-    PageTokens,
-    TableFormerUpdater,
-)
-from docling_eval.docling.utils import (
     docling_version,
     extract_images,
     from_pil_to_base64uri,
     save_shard_to_disk,
 )
+from docling_eval.converters.models.tableformer.tf_model_prediction import (
+    PageTokens,
+    TableFormerUpdater,
+)
+from docling_eval.visualisation.visualisations import save_comparison_html
 
 HTML_EXPORT_LABELS = {
     DocItemLabel.TITLE,
@@ -223,7 +225,8 @@ def create_huggingface_otsl_tableformer_dataset(
                 )
 
             record = {
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: str(ConversionStatus.SUCCESS.value),
                 BenchMarkColumns.DOC_ID: str(os.path.basename(filename)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -234,11 +237,13 @@ def create_huggingface_otsl_tableformer_dataset(
                 BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
                 BenchMarkColumns.GROUNDTRUTH_PAGE_IMAGES: true_page_images,
                 BenchMarkColumns.GROUNDTRUTH_PICTURES: true_pictures,
+                BenchMarkColumns.MODALITIES: [EvaluationModality.TABLE_STRUCTURE],
             }
             records.append(record)
         else:
             record = {
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_TYPE: ConverterTypes.DOCLING,
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: str(ConversionStatus.FAILURE.value),
                 BenchMarkColumns.DOC_ID: str(os.path.basename(filename)),
                 BenchMarkColumns.GROUNDTRUTH: json.dumps(true_doc.export_to_dict()),
@@ -249,6 +254,7 @@ def create_huggingface_otsl_tableformer_dataset(
                 BenchMarkColumns.PREDICTION_PICTURES: pred_pictures,
                 BenchMarkColumns.GROUNDTRUTH_PAGE_IMAGES: true_page_images,
                 BenchMarkColumns.GROUNDTRUTH_PICTURES: true_pictures,
+                BenchMarkColumns.MODALITIES: [EvaluationModality.TABLE_STRUCTURE],
             }
             records.append(record)
