+++ docling_eval/evaluators/base_readingorder_evaluator.py
@@ -16,8 +16,8 @@ from pydantic import BaseModel
 from tqdm import tqdm  # type: ignore
 
 from docling_eval.benchmarks.constants import BenchMarkColumns
-from docling_eval.benchmarks.utils import draw_arrow
 from docling_eval.evaluators.stats import DatasetStatistics, compute_stats
+from docling_eval.visualisation.visualisations import draw_arrow
 
 _log = logging.getLogger(__name__)
