+++ docling_eval/evaluators/bbox_text_evaluator.py
@@ -1,6 +1,6 @@
 import logging
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Dict, List, Optional
 
 import nltk
 from datasets import load_dataset
