+++ docling_eval/evaluators/layout_evaluator.py
@@ -1,6 +1,4 @@
-import copy
 import glob
-import json
 import logging
 from pathlib import Path
 from typing import Dict, List, Tuple
