--- docling_eval/utils/evaluation_error.py
@@ -1,10 +1,0 @@
-from typing import Optional
-
-
-class EvaluationError(Exception):
-    r"""
-    Evaluation error class
-    """
-
-    def __init__(self, msg: Optional[str] = None):
-        Exception.__init__(self, msg)
