+++ docling_eval/visualisation/constants.py
