+++ docs/examples/benchmark_doclaynet_v1.py
@@ -2,8 +2,6 @@ import logging
 import os
 from pathlib import Path
 
-from tabulate import tabulate  # type: ignore
-
 from docling_eval.benchmarks.constants import BenchMarkNames, EvaluationModality
 from docling_eval.benchmarks.doclaynet_v1.create import create_dlnv1_e2e_dataset
 from docling_eval.cli.main import evaluate, visualise
