+++ docs/examples/package_pdfs.py
@@ -9,15 +9,15 @@ from docling_core.types.doc.labels import DocItemLabel
 from tqdm import tqdm  # type: ignore
 
 from docling_eval.benchmarks.constants import BenchMarkColumns
-from docling_eval.benchmarks.utils import draw_clusters_with_reading_order
-from docling_eval.docling.constants import HTML_INSPECTION
-from docling_eval.docling.conversion import create_docling_converter
-from docling_eval.docling.utils import (
+from docling_eval.benchmarks.utils import (
     docling_version,
     from_pil_to_base64,
     get_binary,
     save_shard_to_disk,
 )
+from docling_eval.converters.conversion import create_pdf_docling_converter
+from docling_eval.visualisation.constants import HTML_INSPECTION
+from docling_eval.visualisation.visualisations import draw_clusters_with_reading_order
 
 # Configure logging
 logging.basicConfig(
@@ -133,7 +133,7 @@ def main():
     os.makedirs(viz_dir, exist_ok=True)
 
     # Create Converter
-    doc_converter = create_docling_converter(
+    doc_converter = create_pdf_docling_converter(
         page_image_scale=image_scale, do_ocr=do_ocr
     )
 
@@ -149,7 +149,7 @@ def main():
             pred_doc = conv_results.document
         except:
             record = {
-                BenchMarkColumns.DOCLING_VERSION: docling_version(),
+                BenchMarkColumns.CONVERTER_VERSION: docling_version(),
                 BenchMarkColumns.STATUS: str(ConversionStatus.FAILURE.value),
                 BenchMarkColumns.DOC_ID: str(os.path.basename(pdf_file)),
                 BenchMarkColumns.PREDICTION: json.dumps(None),
@@ -197,7 +197,7 @@ def main():
             fw.write(page)
 
         record = {
-            BenchMarkColumns.DOCLING_VERSION: docling_version(),
+            BenchMarkColumns.CONVERTER_VERSION: docling_version(),
             BenchMarkColumns.STATUS: str(conv_results.status.value),
             BenchMarkColumns.DOC_ID: str(os.path.basename(pdf_file)),
             BenchMarkColumns.PREDICTION: json.dumps(pred_doc.export_to_dict()),
